Update patch set 1
Patch Set 1: Code-Review-1

Aaron fix something similar in https://chromium-review.googlesource.com/416348.

In your case however, there is no GPIO to tell tablet mode.
But on working device, we need to be tablet mode by default and the lid angle calculation is moving it back to clamshell at first calculation: otherwise, when powering init the EC while in tablet mode, the lid angle would be 0 or 360 and the keyboard would be active. We want the keyboard inactive. It is not a problem if lid is cloded: as soon as the lid opens, we will detect clamshell mode and activate the keyboard.

Patch-set: 1
Label: Code-Review=-1