motion: Fix last timestamp calculation
last_collection is used for sensor in forced mode to be sure we are not
calling the read routine too often. We should update last_collection even
when the reading fails, and not on the interrupt path.

For sensors that support interrupt, the new timing diagram is:

              /-------- data rate period ---------\
--------------+------------------------------------+-----------------> t
              |                     /\
              read (sample request)  |
              |                      Interrupt (sample available)
              \/                     |
                sensor

BUG=chrome-os-partner:59423
BRANCH=reef
TEST=Check the ALS read is not called too often even when the sensor
reports an error because the read value as not changed.

Change-Id: I2def7bbd5227cf373c1f613c9b70bc6215861008
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>