Update patch set 5
Patch Set 5: Verified+1

fixed the comment typo.

Guys, I hear your opinions, but I don't see a clear path forward.

Do we agree that suppressing nvmem commits at startup is a good idea? It sure save us to the tune of 250 ms.

If we do - can we go ahead with this patch and then fine tune how we reinstate nvmem commits? The next patch in this stack already introduces a vendor command for that.

Patch-set: 5
Label: Verified=+1