Update patch set 1
Patch Set 1:

> (1 comment)

Yes, to give a bit of context on what I'm doing: I have an ARM board that uses a yocto/oe based workflow. I wanna have ectool on the developer builds of that, so when putting together a recipe I realized it wouldn't specify the correct LDFLAGS if I don't pass in the LDFLAGS.
It basically calls 'make' with HOST_LDFLAGS=foo (for the cross compilation environment)
Do you have a suggestion on how a version in Makefiles.toolchain would look like?

Patch-set: 1