cr50: modify deep sleep behavior
Instead of deciding to enable/disable deep sleep based on some obscure
AP pin state transitions let's use an explicit vendor command for
that.

By default deep sleep would be disabled, only when the AP sends the
vendor command to enable it will it be enabled.

The state would be lost on S3 and deeper suspend, so the AP will have
to re-send the command after resuming from all those states.

Deep sleep will have to remain disabled in S0ix.

The new vendor command allows to send a single byte value to the
target. Accepted values are 0 to disable deep sleep and 1 to enable
deep sleep.

BRANCH=none
BUG=chrome-os-partner:59007

TEST=using the truks_send utility tried sent the following commands
     and observed cr50 behavior (>> denotes cr50 console messages):

  # trunks_send  --raw 80010000000C200000000015
  >> 1913.615628 control_deep_sleep: unexpected input size 0]

  # trunks_send  --raw 80010000000d20000000001500
  >> [1475.486632 control_deep_sleep: deep sleep disabled]

  # trunks_send  --raw 80010000000d20000000001501
  >> [1479.957632 control_deep_sleep: deep sleep enabled]

  # trunks_send  --raw 80010000000d20000000001502
  >> [2037.518020 control_deep_sleep: unexpected command 2]

  # trunks_send  --raw 80010000000e2000000000150000
  >> [2050.082008 control_deep_sleep: unexpected input size 2]

Change-Id: Ia03310516c50cdbcdcd4f4a1b0dda819ff785130