WIP: cr50: battery disconnect in sw
For certain hardware configurations battery disconnect needs to be
handled by the H1 firmware. This patch adds the code for that.

If the charger cable is yanked while Power and Volume Up are pressed,
and stays disconnected for 5 seconds - generate a 5 s pulse on
EC_RST_L and BATT_DISABLE outputs. In fact this pulse would never
complete, as disabling the battery while power is disconnected will
shut down the system completely.

BRANCH=none
BUG=
TEST=

Change-Id: I3b2c1df83522d063e862774fbf4a3a8ac2b7a9b3
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>