rwsig: Put key/signature at the end of RO/RW sections
Current code puts key at the end of the first half of flash, which
may not match the actual RO section size (e.g., it might be PSTATE,
not RO). This makes sure the key to be at the end of RO section, and
signature at the end of RW section, no matter the actual layout
being used.

However, the (deprecated) usbpd1 image type assumes that flash is
equally split between RO and RW, so we do not change that.

BRANCH=none
BUG=chrome-os-partner:61671
TEST=make BOARD=hammer -j && deploy
TEST=make BOARD=hoho -j && \
         futility show --type usbpd1 build/hoho/ec.bin

Change-Id: Ia02e927f9128d6ec3d0b780c28312e0d18835d72