Update patch set 2
Patch Set 2:

The way to use this API is as follows.  It would be nice if the "key fingerprint" is persisted by the NV layer.  This will allow the API to detect when the decryption key mis-matches with the key that was previously used to encrypt the NV data - this will allow NV to be reinitialized in a graceful manner.  The fingerprint is an int.

DCRYPTO_appkey_init(NVMEM, &ctx);
int key_fingerprint = DCRYTPO_appkey_fingerprint(&ctx);
if (!starting_from_scratch && key_fingerprint != stored_key_fingerprint) {
 // Something is wrong here.  The stored key-print differs from the calculated 
 // fingerprint, stored data will not decrypt.
 // Re-initialize NV here?
}

uint8_t *in, *out;
for (block_index = 0; block_index < count; block_index++) {
  DCRYPTO_appkey_cipher_block(&ctx, out, in, block_index);
  out += 16;
  in += 16;
}

Patch-set: 2
Label: Code-Review=0