Update patch set 2
Patch Set 2:

> (2 comments)
 > 
 > Why we haven't see this before is sensors are usually up in S3, so
 > we don't have to run init() again when we resume. We should check
 > the origin of the problem in when called by motion sense task.

But we just moved the sensors to be on in S3. Is this bug specific to older boards where they weren't powered in S3? If the power is what causes this requirement how does S5 work? Do we re-initialize the sensor to normal coming out of S5?

Patch-set: 2