Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> What happens when the AP turns off, cr50 enters deep sleep, and then cr50 wakes up for some reason without the AP turning on? Will cr50 not be able to enter deep sleep until the AP turns on and off again?

this is a good point, with this patch, it will not be able to fall into deep sleep any more.

How about this: deep sleep is enabled by TPM reset (this includes power on).

Receiving TPM_Startup command disables deep sleep until the AP sends the vendor command to re-enable it.

Patch-set: 3