Update patch set 1
Patch Set 1:

> Most other boards are using a 2A limit, Tony suggested we reduce to 1.5A.
> I'm not completely sure to understand the BC1.2 spec, 
> but according to "Figure 4-2 DCP Required Operating Range" I think 2A cutoff might be fine.

a DCP power supply needs to provide at least 1.5A, but it can provide up to 5A.
(and between 2 and 2.5A is very common amongst USB-A chargers)
You need to find the cut-off point to know the max.
Normally it's done with our ramping 
and this value in the code is the max for ramping.

Patch-set: 1