Update patch set 1
Patch Set 1:

We spent a bunch of time looking at different adapters in the samus testing.

There are quite a few adapters at 1.8A (original BC DCP limit) and more at 2A (Samsung popularized) that show as DCP and given we use the ramp code we didn't want to miss out on the extra 2.5W (25%) we could get. We didn't want to set the limit higher than 2A because there were some 1.5A supplies that didn't start to droop until >2.5A and would get very hot in the process. (Most of the 12W adapters we checked advertise as Apple, so we catch them with the other detection -- still need to ramp because some of them lie.)

Given we have not seen issues with the 2A limit, I don't see a need to change it.

Patch-set: 1