Update patch set 2
Patch Set 2:

Thanks for the comments, I had not considered that security implication.

Looking at the pre-existing scheme, one could add a 0xea byte to the RW image somewhere after the real end of image, causing the hash computation to be extended with arbitrary data. Separately, the trailing 0xea byte could be removed, causing the next-to-last 0xea byte to be the end of our hash computation, which could be manipulated through data / instruction selection (much harder to do than just modifying a uint32, but still possible, I think).

Patch-set: 2