common: prepare nvmem for encryption support
This patch mostly is rearranging the code to make it easier to add
encryption layer in the upcoming change.

One substantial change is making sure that in case when NVMEM contents
are corrupted for whatever reason, the initialization function
re-creates a blank NVMEM space; it was just bailing out before not
leaving any initialized partitions behind.

There is no need in two separate tables - one for base absolute
addresses of the NVMEM partitions, and one for their flash memory
offsets, one can be easily derived from the other.

Code erasing the destination partition, calculating the SHA1 of the
new blob and writing it into the flash was separated into own
function.

BRANCH=none
BUG=chrome-os-partner:55331
TEST=make buildall -j still passes (which means NVMEM tests succeed).
     TPM TCG test suite also passes.

Change-Id: I378265d8b49b81398aece2eadf9698abb05caaa1
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>