common: encrypt nvmem
This patch adds an NVMEM encryption layer. To maintain backwards
compatibility the previously unused tag field is now used to determine
if the contents of NVMEM are encrypted. Setting the bit to 1 indicates
unencrypted contents (the legacy case). When the contents are
encrypted, the bit is set to zero.

BRANCH=none
BUG=chromium:
BUG=chrome-os-partner:
TEST=manual

Change-Id: I50b148ac0dc6bc924f4d65c67bc6610100d9dfc0