Update patch set 3
Patch Set 3: Code-Review+1

the description should mention that the default pulse width is being set to 20 ms, and that the option to set it a different interval is added to the command.


And test should demonstrate that the command parameter indeed works (say try 2000 ms and observe it stall for this long).

Patch-set: 3
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1