cr50: x86: only reset the TPM when the AP resumes
The TPM will lock if it sees too many resets in a short time. Right now
the TPM resets whenever the TPM reset signal is asserted and during cr50
init. That means every time Cr50 resumes from deep sleep the TPM is
reset. This change is to reduce the TPM resets to just the necessary
ones on x86. We still need to think through the reset signal behavior on
ARM, so it doesn't change anything on ARM devices.

On x86 we will only reset the TPM when the system is resuming from s3,
s5, or g3, and we will only enable deep sleep in those states. PLT_RST_L
is asserted in S3, s5, and g3, and it is deasserted in s0 and s0ix. This
change switches Reef from using UART to detect the AP state to using
PLT_RST_L. This simplifies the state detection removing the need to
debounce the signal. The PLT_RST_L interrupts will notify the
appropriate CHIPSET hooks, when the state changes. The tpm will reset
only on HOOK_CHIPSET_RESUME.

On ARM Cr50 will still reset the TPM during init and when sys_rst_l is
asserted.

BUG=chrome-os-partner:60555
BUG=chrome-os-partner:59007
BRANCH=none
TEST=manual
	On Reef

	make sure cr50 does not receive a sys_rst_asserted interrupt
	even when sys_rst_l is asserted

	Put the system into different power states

	verify Cr50 only enters deep sleep in s3, s5, and g3.

	put the system into s0. wait until cr50 enters regular sleep.

	from the ec run 'apreset' verify cr50 wakes up on the falling
	edge of plt_rst_l, sees that plt_rst_l is asserted and sees the
	deassertion.

	Make sure the tpm is reset on that pulse

	Verify Gru still works

Change-Id: Icfdaf31533719cbce5838c99c8f41658a6d94e28
Signed-off-by: Mary Ruthven <mruthven@chromium.org>