Update patch set 3
Patch Set 3:

(1 comment)

Thinking more, I'm not sure we want to land this change (except the assert). We probably want an upper bound for the size of the buffer as well as a lower bound -- if the buffer is too large, we may spend too much time hashing and starve out other deferred tasks.

Patch-set: 3