Update patch set 4
Patch Set 4: Code-Review+2

This is fine for our FW branch here, we're not likely to have a shared mem. buffer that is "too large" since we're fighting space constraints on this part. But if we bring the CL to TOT, we should probably add an upper bound for CHUNK_SIZE.

Patch-set: 4
Reviewer: Gerrit User 1001696 <1001696@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2