Use lid angle to check lid open state supplementarily
NOT_READY_FOR_REVIEW

When two chromebooks are stacked, the magnet in the lid of the
bottom device can weaken the magnetic field created by
the magnet in the lid of the top device, causing the lid open
event to trigger.

This patch makes the hook task double check if the lid is really
open or not by reading the last lid angle. If it is greater than
lid_open_angle, the lid open event triggers. Otherwise, the lid
open event will be ignored.

This should not influence if the device has no lid angle calculation
capability (CONFIG_LID_ANGLE is undef).

BUG=chrome-os-partner:61971
BRANCH=none
TEST=make buidlall. With two Reefs stacked, the bottom lid magnet
does not wake up the device on top from S3. A Reef wakes up from S3
when the lid is opened.

Change-Id: I9309807d34fa46d640ed5e801ab73687d71c8b8e
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>