nvmem: encrypt contents using crypto api
This patch makes incompatible changes to the nvmem layout: the header
is increased to accommodate a 16 byte sha ans a 16 byte padding for
future extensions.

The layout version field is also introduced to make it easier to track
changes in the future. When calculating SHA the entire partition above
the SHA field is processed. Encryption covers everything above the
header.

Introducing encryption makes it impossible to use flash contents
directly for read and compare operations.

BRANCH=none
BUG=chrome-os-partner:62260
TEST=make buildall -j passes
     tcg tests pass
     it is possible to corp enroll on reef

Change-Id: I50b148ac0dc6bc924f4d65c67bc6610100d9dfc0