lid_switch: Prevent false lid open from waking-up device
When a device with the lid closed is stacked on top of another,
the bottom lid magnet may create magnetic field strong enough to
cancel the top lid magnetic field, thus the device wakes up
unintentionally by a (false) lid open event. This is
indistinguishable from a real lid open event (unless the bottom
lid magnet is so strong that it also triggers TABLET_MODE_L) by
lid switch itself.

To avoid this, we read a lid angle to see the lid is really open.

lid_angle_closed_low has to be smaller than the angle at which
LID_OPEN triggers and larger than the angle of the closed lid.

BUG=b:35586669
BRANCH=none
TEST=make buildall. Place an Electro in S3 on another. Ensure
false lid open is detected and it doesn't wake up. Ensure
Electro wakes up normally when lid is opened.

Change-Id: I665e685262dcceaadd182e87b4a993aeee6dbfaf
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>