lid_switch: Prevent false lid open from waking-up device
When a device is stacked on top of another, the bottom lid magnet
may create magnetic field strong enough to cancel the top lid
magnetic field, which is sustaining lid close state. This is
indistinguishable from a real lid open event (unless the bottom
lid magnet is so strong that it also triggers TABLET_MODE_L).

When this happens, the system wakes up from S3 or S5 even with
the lid closed. To avoid it, we read a lid angle and make
sure the lid is really open.

lid_angle_calculation_delay_msec is necessary to avoid slowly opened
lid returns a small angle (thus we mistakenly judges it as a false
lid open). The longer the dealy is the more reliably we can judge
the lid open but of course a longer delay causes a slower wake-up.

BUG=chrome-os-partner:61971
BRANCH=none
TEST=make buildall. Place an Electro in S3 on another. Make sure false
lid open is detected and it doesn't wake up. Ensure Electro
wakes up or boot normally in other situations.

Change-Id: I665e685262dcceaadd182e87b4a993aeee6dbfaf
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>