Update patch set 2
Patch Set 2:

Maybe.

I think we need to discuss this in the context of an overall revisiting of how Haven managed write protect, not as a one-off before then, and not without involving the security team.  

Some of the changes up in the air to support the test lab flow may interact poorly with weakening the unlock process.

Patch-set: 2