Makefile.rules: Clean ccache cache for 'clobber'.
In the chroot, we use ccache to speed up compilation.  However, it's
possible for this cache to become corrupt which could lead to slower
builds.  In order to keep things fresh, 'make clobber' should clean up
that ccache cache as well.  This commit makes it so.

Note, this is NOT *clearing* the cache entirely.

BUG=chromium:681961
BRANCH=None
TEST=make clobber; verify that ccache is cleaned.

Change-Id: I8ccd7df4858c445485ed740843f3126af0abd33a
Signed-off-by: Aseda Aboagye <aaboagye@google.com>