motion_sense: Do not switch rate if sensor rails are off
During shutdown sequence some PMICs (Ex:TPS65094x) turn off the U-rail
even before PMIC is disabled. If we try to access the sensor which is
defined active in S3 but it is powered-off we get I2C errors. To avoid
this issue, do not switch the sensor rate if in S3 & S3 configs are
same as in S0.

BUG=chrome-os-partner:61489
BRANCH=none
TEST=Manually tested on Electro. Put the device in S5 and observed
     no I2C errors printed on EC the console.

Change-Id: I40b75ebbf0d05fafffdfd535962423c3960fafbf
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>