Update patch set 1
Patch Set 1:

(1 comment)

> Vadim - have you thought about changing your scheme such that you
 > don't have to read the entire contents of the flash but only read
 > 1K at a time and decrypt what is needed maybe?  Maybe reading only
 > a subset of the flash at a time would be fast enough that you don't
 > have to even build a cache in memory?  This would also reduce your
 > memory footprint requirements

Actually a dcrypto speed up patch has been posted by ngm@, it reduces the encryption/decryption time to less than 3 ms.

I don't think it is possible to decrypt part of the crypto text starting at some point into the text. It is possible to decrypt part of the text starting from the beginning, but if clear text changes the entire crypto text will have to be changed.

A bigger issue is that the underlying nvmem architecture requires writing the entire blob (16K at this time). We have been thinking about limiting this to one flash block, but it is not obvious how to do this and satisfy the TPM use pattern, so this optimization has been put off.

With the next patch in the stack the nvmem size reduces from 16K to 12K (pending apronin@'s verification that this is enough)

Patch-set: 1