Update patch set 2
Patch Set 2:

the idea is that there is a crypto API, defined in include/crypto_api.h.

Its implementation comes from the chip (in this case DCRYPTO_xxxx functions).

If cr50 board is moved to a new chip, that chip will have to provide implementation for this API. If the cr50 board does not use the new chip - it does not include the new chip's implementation of crypto API. Does this make sense?

Patch-set: 2