motion_sense: Defer switching the sensors rate if in S3
During shutdown sequence some PMICs (ex:TPS65094x) turn off the sensor
rails asynchronously to the EC. If we access the sensors when the sensor
power rails are off we get I2C errors. To avoid this issue, defer switching
the sensors rate if in S3. By the time deferred function is serviced if
the chipset is in S5 we can back out from switching the sensor rate.

BUG=chrome-os-partner:61489
BRANCH=none
TEST=Manually tested on Electro. Put the device in S5 and observed
     no I2C errors printed on EC the console.

Change-Id: I40b75ebbf0d05fafffdfd535962423c3960fafbf
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>