Update patch set 2
Patch Set 2:

> I don't think it is possible to decrypt part of the crypto text
 > starting at some point into the text.

Actually this is possible... it's called "random access"
(hence the choice of a counter mode cipher).

In the next revision to nvmem architecture, decryption can be
done inline directly into a user buffer (on whatever granularity/
offset is necessary), thus avoiding the need for a read-side cache.

The write side cache will still have to exist as is the case with
unencrypted nvmem.  Though it is as possible to make writes granular by
using per page salts (or generation numbers etc.)

Patch-set: 2