caroline: do not boot with 15W in read-only firmware
Increase the minimum charger power to boot with the battery ensuring a
minimum buffer to 25W.
And add a new CONFIG_CHARGER_LIMIT_POWER_ENFORCE_RO_THRESH parameter to
even forbid it to boot in RO to do the EC software-sync if this 25W
threshold is not reached.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=glados
BUG=chrome-os-partner:58657
TEST=Discharge caroline battery down to less than 7V.
Plug AC and try repeatedly to boot the system.
See it won't boot until battery state of charge is 1%.

Change-Id: I0ff0dc9bced1eb9309f13b35943ffb675aebafae