Update patch set 6
Patch Set 6:

(1 comment)

Last thing, unless I missed commits_skipped being used in some manner.

Also, I think we should file a bug on re-evaluating the role of nvmem and the semantics it's providing to users. It's pitched as multiple task reading/writing, but there's some issues lurking that may cause confusion if people consider it to be a truly generic library when multiple tasks are utilizing it.

Patch-set: 6
Label: Code-Review=0