Update patch set 3
Patch Set 3: Code-Review+2

DCRYPTO_SHA1_hash() is what's in chip/g/

app_compute_hash() is not chip-specific. All it's doing is putting a wrapper around the existing chip-specific function and changing the name.

app_compute_hash() should be in either board/cr50 or in common/ (guarded with CONFIG_DCRYPTO). It would work just as well as a pure-software implementation, and should probably be renamed compute_sha1(), since that's what it does. 

Actually, we already HAVE generic sha1 functions, in include/sha1.h:

  sha1_init(), sha1_update(), sha1_final()

The software versions are used with CONFIG_SHA1 and the HW versions are used with CONFIG_DCRYPTO.

The app_compute_hash() function shouldn't exist at all. If it must, make it file-scoped.

Patch-set: 3