Update patch set 3
Patch Set 3: Code-Review-2

Removing my +2, replacing with a -2.

This CL should be abandoned, and replaced with one that leaves the decision of how common/nvmem.c should compute a hash up to common/nvmem.c. We already have APIs for the standard crypto functions, which may be done in software or chip-specific hardware. We don't need new names for specific calls to those functions, just because nvmem wants a hashing function.

Patch-set: 3