Update patch set 3
Patch Set 3:

just to reiterate: nvmem should have access to a system wide abstraction which allows to calculate a hash.

that abstraction should be fulfilled by the chip, so the implementation belongs in the chip space.

Guys, if this is about replacing sha1 with sha256 - it is ok with me. I was against mixing it up with other changes, but on its own -fine with me.

Patch-set: 3