g: reintroduce the cipher console command
The cipher command implementation had to change to account for reduced
heap size and to improve coverage.

Note that the goal of this command is not to verify correctness of the
encryption services provided by the dcrypto, but to allow to exercise
the crypto engine on multiple passes, each time using the same clear
text but different initialization vector.

BRANCH=none
BUG=chrome-os-partner:62260
TEST=ran cipher command on a few devices:
  > cip
  Will wait up to 4074 ms
  running 1000 iterations
  blob size 7111 at 1e020
  original data  8f3d99fbfcbd26dd0c4d8dc444d106ee
  hashed data    826a4e9b04d214fbbd5fbf4e0fba8068
  Encryption results: min 1128 us, max 1456 us, average 1180 us
  Decryption results: min 1124 us, max 7348 us, average 1193 us
  >

Change-Id: Idf72b355dce0f288d4a3d8a065bc08eb9c4f6bc3
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>