g: use H1_DEVIDS as a trigger to use the fob for signing
The CR50_DEV make variable is loaded: it enables debug extensions in
the produced cr50 image AND requires that the image is signed with the
key from the signing fob.

In fact these are two independent requirements: it is possible to use
an image built with CR50_DEV set on a dev H1 which does not require
fob signing.

A proper indication of need to use the signing fob would be the fact
that H1_DEVIDS is defined, as it means a that node locked image is
being produced.

BRANCH=none
BUG=chrome-os-partner:62457
TEST=not fully tested yet

Change-Id: Ied65a0bc50926aa5b6fa65e51805c2368522dcf2
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>