ANX74xx: Mimic the cable detect interrupt
eMark cable is not detected in low power mode if not powered initially
hence we don't get the cable detect interrupt when the cable is powered.
To avoid this issue, before entering low power mode get the VBUS status
from the charger and override the low power mode to normal mode to mimic
the cable detect interrupt.

BUG=chrome-os-partner:62550
BRANCH=none
TEST=Manually tested on Electro.
     Connect the eMark cable to Electro, power on the external charger,
     and observed reef can charge.

Change-Id: I6672ecdf4e58b8da68bfa1b5ae23a75a43858ae1
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>